python-project/spade-tools/requirements.txt
redis==3.5.3
langchain>=0.1.0
openai>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
lxml>=4.9.0
pyyaml>=6.0.1
langchain-community~=0.3.15
playwright~=1.49.1
setuptools~=75.1.0
langchain-openai~=0.3.2
python-docx>=2.8.1
